Bug 1000: Fiexed a lot of small issues during the MAC busy week. OTDB: getTreeGroup(3) is now ordered on stoptime iso starttime. GCFPVDynArr: wrote operator==() 2 new errors: CT_RESULT_NO_PARSET and CT_RESULT_UNKNOWN_OBSERVATION ChildControl: copying parsetfile to remote system is done asynchronous with MAC_SCP_TIMEOUT secs timeout. 'Active' state renamed to 'Running' SoftwareMonitor extended with list of restartable programs. StationControl can handle startcommand when parset is not their. MACScheduler: Observationslist are only written when they are changed.    Alarm is raised in the Navigator when observation is aborted. ObservationControl: runState field is kept up to date with state of observation.

+++ branches/MAC-20101206-1_0-release-branch/MAC/APL/PIC/RSP_Driver/src/SetHBACmd.cc
@@ -112,5 +112,5 @@ bool SetHBACmd::validate() const
   return ((m_event->rcumask.count() <= (unsigned int)StationSettings::instance()->nrRcus())
 	  	&& (2 == m_event->settings().dimensions())
 		&& (m_event->rcumask.count() == (unsigned int)m_event->settings().extent(firstDim))		// check number off selected rcus
-		&& (N_HBA_ELEM_PER_TILE == (unsigned int)m_event->settings().extent(secondDim))); // check number of elements
+		&& (N_HBA_ELEM_PER_TILE == m_event->settings().extent(secondDim))); // check number of elements
 }
